LRSD-6817 rename
