+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-etc-spring-boot/LCP.json
