+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-etc-spring-boot/src/main/java/com/liferay/learn/LearnSpringBootApplication.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.lms;
+package com.liferay.learn;
 
 import com.liferay.client.extension.util.spring.boot.ClientExtensionUtilSpringBootComponentScan;
 
@@ -16,10 +16,10 @@ import org.springframework.context.annotation.Import;
  */
 @Import(ClientExtensionUtilSpringBootComponentScan.class)
 @SpringBootApplication
-public class LMSSpringBootApplication {
+public class LearnSpringBootApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LMSSpringBootApplication.class, args);
+		SpringApplication.run(LearnSpringBootApplication.class, args);
 	}
 
 }
