+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-etc-spring-boot/src/main/java/com/liferay/learn/ReadyRestController.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.lms;
+package com.liferay.learn;
 
 import com.liferay.client.extension.util.spring.boot.BaseRestController;
