+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-etc-spring-boot/src/main/resources/application-default.properties
@@ -2,7 +2,7 @@
 # Application
 #
 
-liferay.lms.dxp.site.group.id=
+liferay.learn.dxp.site.group.id=
 
 #
 # LXC
@@ -16,7 +16,7 @@ com.liferay.lxc.dxp.server.protocol=http
 # OAuth
 #
 
-liferay.oauth.application.external.reference.codes=liferay-lms-etc-spring-boot-oauth-application-user-agent
+liferay.oauth.application.external.reference.codes=liferay-learn-etc-spring-boot-oauth-application-user-agent
 liferay.oauth.urls.excludes=/ready
 
 #
