+++ workspaces/liferay-learn-workspace/client-extensions/liferay-learn-etc-spring-boot/src/main/resources/application.properties
