--- workspaces/liferay-learn-workspace/client-extensions/liferay-lms-etc-spring-boot/client-extension.yaml
@@ -1,16 +1,0 @@
-assemble:
-    - fromTask: bootJar
-liferay-lms-etc-spring-boot-oauth-application-user-agent:
-    .serviceAddress: localhost:58081
-    .serviceScheme: http
-    name: Liferay LMS Etc Spring Boot OAuth Application User Agent
-    scopes:
-        - c_course.everything
-        - c_lesson.everything
-        - c_module.everything
-    type: oAuthApplicationUserAgent
-liferay-lms-etc-spring-boot-object-action-course-duration:
-    name: Liferay LMS Etc Spring Boot Object Action Course Duration
-    oAuth2ApplicationExternalReferenceCode: liferay-lms-etc-spring-boot-oauth-application-user-agent
-    resourcePath: /object/action/course-duration
-    type: objectAction
